Boton Actualizar y compartir

+++ app/src/main/java/com/example/pm2e10087/ActivityList.java
@@ -33,8 +33,8 @@ public class ActivityList extends AppCompatActivity {
     ArrayList<Contactos> lista;
     List<Contactos> lis;
     ArrayList<String> listaConcatenada;
-
-    Button btnEliminar;
+    String id = "";
+    Button btnEliminar, btnActualizar, btnCompartir;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -44,7 +44,11 @@ public class ActivityList extends AppCompatActivity {
         conexion = new SQLiteConexion(this, Transacciones.NameDatabase, null, 1);
         listContactos = (ListView) findViewById(R.id.listContactos);
         btnEliminar = findViewById(R.id.btnEliminar);
+        btnActualizar = findViewById(R.id.btnActualizar);
+        btnCompartir = findViewById(R.id.btnCompartir);
 
+
+        //llenar lista
         GetListContactos();
 
         ArrayAdapter adp = new ArrayAdapter(this, android.R.layout.simple_list_item_1, listaConcatenada);
@@ -55,14 +59,34 @@ public class ActivityList extends AppCompatActivity {
             public void onItemClick(AdapterView<?> adapterView, View view, int i, long l) {
                 //Toast.makeText(ActivityList.this, listaConcatenada.get(i), Toast.LENGTH_SHORT ).show();
                 Log.i("Seleccion id", listaConcatenada.get(i));
-
-
+                id = lista.get(i).getId().toString();
+                //Boton Eliminar Registro
                 btnEliminar.setOnClickListener(new View.OnClickListener() {
                     @Override
                     public void onClick(View view) {
-                        String id = lista.get(i).getId().toString();
-                        mostrarAlerta("¿Desea eliminar el numero de contacto "+lista.get(i).gettelefono(),id);
+                        id = lista.get(i).getId().toString();
+                        mostrarAlerta("¿Desea eliminar el numero de contacto "+lista.get(i).gettelefono(),id,"D");
                         Log.i("Eliminar id", ""+lista.get(i).getId().toString());
+                    }
+                });
+
+                //Boton Actualizar Registro
+                btnActualizar.setOnClickListener(new View.OnClickListener() {
+                    @Override
+                    public void onClick(View view) {
+                        id = lista.get(i).getId().toString();
+                        mostrarAlerta("¿Desea actualizar el numero de contacto "+lista.get(i).gettelefono(),id, "A");
+                        Log.i("Actualizar id", ""+lista.get(i).getId().toString());
+
+
+                    }
+                });
+
+                //Boton Compartir
+                btnCompartir.setOnClickListener(new View.OnClickListener() {
+                    @Override
+                    public void onClick(View view) {
+                        compartir(lista.get(i).getNombre().toString(), lista.get(i).gettelefono().toString());
                     }
                 });
             }
@@ -103,26 +127,45 @@ public class ActivityList extends AppCompatActivity {
         }
     }
 
-    private void mostrarAlerta(String mensaje, String id) {
+    private void mostrarAlerta(String mensaje, String id, String tipo) {
+
         AlertDialog.Builder builder = new AlertDialog.Builder(ActivityList.this);
         builder.setMessage(mensaje)
                 .setTitle("Confirmación")
                 .setPositiveButton("Sí", new DialogInterface.OnClickListener() {
                     @Override
                     public void onClick(DialogInterface dialogInterface, int i) {
-                        eliminarRegistro(id);
+                      if (tipo.equalsIgnoreCase("D")){ //Delete
+                          eliminarRegistro(id);
+                      }
+                        if (tipo.equalsIgnoreCase("A")){ //Actualizar
+                            actualizarRegistro(id);
+                        }
                     }
                 }
                 ).setNegativeButton("No", new DialogInterface.OnClickListener() {
                             @Override
                             public void onClick(DialogInterface dialogInterface, int i) {
                                 dialogInterface.cancel();
+
                             }
                         }
                 );
 
         AlertDialog dialog = builder.create();
         dialog.show();
+    }
+
+    private void actualizarRegistro(String id) {
+        try {
+            onBackPressed();
+        } catch (Throwable e) {
+            e.printStackTrace();
+        }
+        Intent intent = new Intent(ActivityList.this, MainActivity.class);
+        intent.putExtra("id", id);
+        startActivity(intent);
+
     }
 
     private void eliminarRegistro(String id) {
@@ -140,9 +183,47 @@ public class ActivityList extends AppCompatActivity {
         }
 
         db.close();
-
+        try {
+            onBackPressed();
+        } catch (Throwable e) {
+            e.printStackTrace();
+        }
         //Volver a Activity principal
         Intent intent = new Intent(ActivityList.this, MainActivity.class);
         startActivity(intent);
+    }
+
+    public void verificaSeleccion(View view){
+
+        if (id.toString().equalsIgnoreCase("")){
+            mostrarAlertaOK("Debe seleccionar un registro y luego presionar el boton");
+        }
+    }
+
+    private void mostrarAlertaOK(String mensaje) {
+        AlertDialog.Builder builder = new AlertDialog.Builder(ActivityList.this);
+        builder.setMessage(mensaje)
+                .setTitle("Error")
+                .setPositiveButton("Ok", new DialogInterface.OnClickListener() {
+                    @Override
+                    public void onClick(DialogInterface dialogInterface, int i) {
+                        dialogInterface.cancel();
+                    }
+                });
+
+        AlertDialog dialog = builder.create();
+        dialog.show();
+    }
+
+    private void compartir(String nombre, String telefono){
+        if (id != null){
+            Intent intentShare;
+            intentShare = new Intent(Intent.ACTION_SEND);
+            intentShare.setType("text/pain");
+            intentShare.putExtra(android.content.Intent.EXTRA_SUBJECT,"Contacto");
+            intentShare.putExtra(android.content.Intent.EXTRA_TEXT,"Nombre: "+nombre+"\nTelefono: "+telefono);
+            startActivity(Intent.createChooser(intentShare,"Compartir con"));
+        }
+
     }
 }
